+++ src/app/app.module.ts
@@ -14,32 +14,57 @@
  *   along with this program, in the LICENSE.md file in the root project directory.
  *   If not, see https://www.gnu.org/licenses/lgpl-3.0.txt
  */
-
-import { HttpClientModule, HTTP_INTERCEPTORS } from "@angular/common/http";
-import { NgModule } from "@angular/core";
-import { BrowserModule } from "@angular/platform-browser";
-import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
-import { AppRoutingModule } from "./app-routing.module";
-import { AppComponent } from "./app.component";
-import { AppLayoutComponent } from "./components/app-layout/app-layout.component";
-import { AuthCallbackComponent } from "./components/auth-callback/auth-callback.component";
-import { HomeComponent } from "./components/home/home.component";
-import { TokenInterceptor } from "./interceptors/auth.interceptor";
-import { TopHighlightCardsComponent } from "./modules/highlight/top-highlight-cards/top-highlight-cards.component";
-import { SharedModule } from "./modules/shared/shared.module";
+import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
+import { NgModule, ErrorHandler } from '@angular/core';
+import { BrowserModule } from '@angular/platform-browser';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { AppRoutingModule } from './app-routing.module';
+import { AppComponent } from './app.component';
+import { AppLayoutComponent } from './components/app-layout/app-layout.component';
+import { AuthCallbackComponent } from './components/auth-callback/auth-callback.component';
+import { HomeComponent } from './components/home/home.component';
+import { TokenInterceptor } from './interceptors/auth.interceptor';
+import { TopHighlightCardsComponent } from './modules/highlight/top-highlight-cards/top-highlight-cards.component';
+import { SharedModule } from './modules/shared/shared.module';
 import { BsDropdownModule } from 'ngx-bootstrap/dropdown';
 import { FooterComponent } from './components/footer/footer.component';
+import { HttpErrorInterceptor } from './interceptors/http.interceptor';
+import { SentryErrorHandler } from './error-handler/sentry.error-handler';
+import { errorHandlerFactory } from './error-handler/error-handler-factory';
 
 @NgModule({
-  declarations: [AppComponent, AppLayoutComponent, HomeComponent, AuthCallbackComponent, TopHighlightCardsComponent, FooterComponent],
-  imports: [BsDropdownModule.forRoot(), BrowserModule, AppRoutingModule, BrowserAnimationsModule, HttpClientModule, SharedModule],
+  declarations: [
+    AppComponent,
+    AppLayoutComponent,
+    HomeComponent,
+    AuthCallbackComponent,
+    TopHighlightCardsComponent,
+    FooterComponent
+  ],
+  imports: [
+    BsDropdownModule.forRoot(),
+    BrowserModule,
+    AppRoutingModule,
+    BrowserAnimationsModule,
+    HttpClientModule,
+    SharedModule
+  ],
   providers: [
     {
       provide: HTTP_INTERCEPTORS,
       useClass: TokenInterceptor,
       multi: true,
     },
+    {
+      provide: HTTP_INTERCEPTORS,
+      useClass: HttpErrorInterceptor,
+      multi: true,
+    },
+    {
+      provide: ErrorHandler,
+      useFactory: errorHandlerFactory
+    }
   ],
   bootstrap: [AppComponent],
 })
-export class AppModule {}
+export class AppModule { }