+++ src/app/error-handler/error-handler-factory.ts
@@ -1,4 +1,5 @@
 /*
+ *
  *  Digital Excellence Copyright (C) 2020 Brend Smits
  *
  *   This program is free software: you can redistribute it and/or modify
@@ -13,16 +14,18 @@
  *   You can find a copy of the GNU Lesser General Public License
  *   along with this program, in the LICENSE.md file in the root project directory.
  *   If not, see https://www.gnu.org/licenses/lgpl-3.0.txt
+ *
  */
+import { SentryErrorHandler } from './sentry.error-handler';
+import { ErrorHandler } from '@angular/core';
+import { environment } from 'src/environments/environment';
 
-export const environment = {
-  production: true,
-  identityServerUrl: 'https://identity.staging.dex.software/',
-  apiUrl: 'https://api.staging.dex.software/',
-  frontendUrl: 'https://staging.dex.software/',
-  identityCallbackUrl: 'https://staging.dex.software/',
-  identityClientId: 'dex-frontend',
-  identityRedirectUri: 'https://identity.staging.dex.software/auth-callback',
-  identityLogoutRedirectUri: 'https://staging.dex.software/',
-  identitySilentRedirectUri: 'https://staging.dex.software/silent-refresh.html',
-};
+/**
+ * Factory for selecting the right error handler based on the environment.
+ */
+export function errorHandlerFactory() {
+    if (environment.production) {
+        return new SentryErrorHandler();
+    }
+    return new ErrorHandler();
+}